Add useTeams hook and simplify RecTile component
Created a new `useTeams` hook to handle team splitting and mapping logic separately. Refactored the `RecTile` component to utilize this hook, resulting in cleaner and more maintainable code. Removed the redundant `showMap` prop from `RecTile` in `top-recorded-games`.
